requirement.txt
afinn==0.1
anyascii==0.3.0
anyio==3.2.1
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.2.10
async-generator==1.10
attrs==19.3.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bcrypt==3.1.7
beautifulsoup4==4.9.3
bleach==3.3.1
blessings==1.7
blinker==1.4
blis==0.7.4
bootstrap4==0.1.0
bqplot==0.12.30
branca==0.4.2
Brlapi==0.7.0
CacheControl==0.12.6
cachetools==4.2.2
catalogue==2.0.6
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
cliapp==1.20180812.1
click==7.1.2
cmdtest==0.32+git
colorama==0.4.3
colour==0.1.5
command-not-found==0.3
contextlib2==0.6.0
crispy-forms-materialize==0.2
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.5
dbfpy==2.3.1
dbus-python==1.2.16
debugpy==1.3.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.0
django-allauth==0.44.0
django-appconf==1.0.4
django-bootstrap-datepicker-plus==3.0.5
django-cors-headers==3.7.0
django-cors-headers-multi==1.2.0
django-crispy-forms==1.12.0
django-cryptography==1.0
django-datetimepicker==3.14
django-environ==0.4.5
django-fernet-fields==0.6
django-filter==2.4.0
django-import-export==2.5.0
django-modelcluster==5.1
django-multiselectfield==0.1.12
django-pwa==1.0.10
django-report-builder==6.4.2
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.12.1
draftjs-exporter==2.1.7
duplicity==0.8.12.0
earthengine-api==0.1.273
easygui==0.98.2
ee==0.2
entrypoints==0.3
et-xmlfile==1.0.1
fasteners==0.14.1
feedfinder2==0.0.4
feedparser==6.0.8
ffmpeg-python==0.2.0
filelock==3.0.12
Flask==2.0.2
folium==0.12.1
future==0.18.2
futures==3.1.1
GDAL==3.0.4
geeadd==0.5.5
geemap==0.8.18
gensim==4.1.2
geocoder==1.38.1
geojson==2.5.0
google-api-core==1.30.0
google-api-python-client==1.12.8
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.1
google-cloud-storage==1.41.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
googledrivedownloader==0.4
googletrans==3.1.0a0
greenlet==1.1.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.10.1
html5lib==1.0.1
httpcore==0.9.1
httplib2==0.19.1
httplib2shim==0.0.3
httpx==0.13.3
hyperframe==5.2.0
icu==0.0.1
idna==2.8
ImageHash==4.2.0
importlib-metadata==1.5.0
indic-transliteration==2.3.2
ipaddr==2.2.0
ipyevents==2.0.1
ipyfilechooser==0.4.4
ipykernel==6.0.2
ipyleaflet==0.14.0
ipynb-py-convert==0.4.6
ipython==7.25.0
ipython-genutils==0.2.0
ipytree==0.2.1
ipywidgets==7.6.3
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.18.0
jieba3k==0.35.1
Jinja2==3.0.2
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.6.3
jupyter-server==1.9.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kazam==1.4.5
keyring==18.0.1
kiwisolver==1.0.1
l18n==2018.5
langid==1.1.6
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
logzero==1.7.0
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib==3.1.2
matplotlib-inline==0.1.2
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
Morfessor==2.0.6
msgpack==0.6.2
mss==6.1.0
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.0.4
nest-asyncio==1.5.1
netifaces==0.10.4
newspaper3k==0.2.8
nltk==3.6.3
notebook==6.4.0
numpy==1.17.4
oauth2client==4.1.3
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.5
OWSLib==0.19.1
packaging==20.3
pandas==1.2.5
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.2
pathy==0.6.0
pep517==0.8.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.4.1
polyglot==16.7.4
preshed==3.0.5
progress==1.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psycopg2==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycld2==0.41
pycparser==2.20
PyCRS==1.0.2
pycups==1.9.73
pydantic==1.8.2
Pygments==2.9.0
PyGObject==3.36.0
PyICU==2.7.4
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyproj==2.5.0
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
pyshp==2.1.3
python-apt==2.0.0+ubuntu0.20.4.4
python-box==5.3.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.4
python3-openid==3.2.0
pytoml==0.1.21
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
ratelim==0.1.6
readability-lxml==0.8.1
regex==2020.10.11
reportlab==3.5.34
requests==2.22.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
retrying==1.3.3
rfc3986==1.5.0
router==0.1
rsa==4.7.2
sankee==0.0.5
scipy==1.6.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.7.1
sgmllib3k==1.0.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smart-open==5.2.1
sniffio==1.2.0
soupsieve==1.9.5
spacy==3.1.3
spacy-legacy==3.0.8
SQLAlchemy==1.4.23
sqlparse==0.2.4
srsly==2.4.1
ssh-import-id==5.10
systemd-python==234
tablib==2.0.0
telepath==0.2
terminado==0.10.1
testpath==0.5.0
textblob==0.15.3
thinc==8.0.10
tinysegmenter==0.3
tldextract==3.1.2
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
traittypes==0.2.1
ttystatus==0.38
typer==0.3.2
typing-extensions==3.10.0.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
voila==0.2.10
wadllib==1.3.3
wagtail==2.14.1
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.2
whitebox==1.5.1
whiteboxgui==0.3.0
widgetsnbextension==3.5.1
Willow==1.4
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==1.3.6
xlwt==1.3.0
zipp==1.0.0
